NRFFOETT-1700: Correct the type of buffer size in ADC driver
It needs to be 'uint16_t' to be in accordance with the parameter taken by
'nrfx_adc_buffer_convert'.

+++ drivers/src/nrfx_adc.c
@@ -17,8 +17,8 @@ typedef struct
     nrfx_adc_channel_t     * p_head;
     nrfx_adc_channel_t     * p_current_conv;
     nrf_adc_value_t        * p_buffer;
-    uint8_t                  size;
-    uint8_t                  idx;
+    uint16_t                 size;
+    uint16_t                 idx;
     nrfx_drv_state_t         state;
 } adc_cb_t;
